# src/main/java/com/postech/fastfood/infrastructure/controller/CustomerController.java
package com.postech.fastfood.infrastructure.controller;

import com.postech.fastfood.application.gateways.LoggerPort;
import com.postech.fastfood.application.mapper.CustomerMapper;
import com.postech.fastfood.domain.User;
import com.postech.fastfood.infrastructure.controller.dto.request.CustomerByEmailRequest;
import com.postech.fastfood.infrastructure.controller.dto.request.CustomerCpfRequest;
import com.postech.fastfood.infrastructure.controller.dto.request.CustomerEmailRequest;
import com.postech.fastfood.application.usecases.interfaces.FindUserByCpfUseCase;
import com.postech.fastfood.application.usecases.interfaces.FindUserByEmailUseCase;
import com.postech.fastfood.application.usecases.interfaces.customer.CreateCustomerWithCpfUseCase;
import com.postech.fastfood.application.usecases.interfaces.customer.CreateCustomerWithNameAndEmailUseCase;
import jakarta.validation.Valid;
import org.hibernate.validator.constraints.br.CPF;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/customer")
public class CustomerController {

    private final CreateCustomerWithCpfUseCase createCustomerWithCpfUseCase;
    private final CreateCustomerWithNameAndEmailUseCase createCustomerWithNameAndEmailUseCase;
    private final FindUserByEmailUseCase findUserByEmail;
    private final FindUserByCpfUseCase findUserByCpfUseCase;
    private final LoggerPort logger;

    public CustomerController(
            CreateCustomerWithCpfUseCase createCustomerWithCpfUseCase,
            CreateCustomerWithNameAndEmailUseCase createCustomerWithNameAndEmailUseCase,
            FindUserByEmailUseCase findCustomerByEmail,
            FindUserByCpfUseCase findUserByCpfUseCase,
            LoggerPort logger) {
        this.createCustomerWithCpfUseCase = createCustomerWithCpfUseCase;
        this.createCustomerWithNameAndEmailUseCase = createCustomerWithNameAndEmailUseCase;
        this.findUserByEmail = findCustomerByEmail;
        this.findUserByCpfUseCase = findUserByCpfUseCase;
        this.logger = logger;
    }

    @PostMapping("/cpf")
    public ResponseEntity<User> createCustomerWithCpf(@RequestBody @Valid CustomerCpfRequest customerCpfRequest) {
        logger.info("[Customer] Iniciando criação de cliente por CPF");
        final User customerCpf = CustomerMapper.toDomain(customerCpfRequest);
        final User customerSaved = this.createCustomerWithCpfUseCase.execute(customerCpf);
        logger.info("[Customer] Cliente criado com sucesso por CPF: id={}", customerSaved.getId());
        return ResponseEntity.status(HttpStatus.CREATED).body(customerSaved);
    }

    @PostMapping("/email")
    public ResponseEntity<User> createCustomerWithEmailAndName(@RequestBody @Valid CustomerEmailRequest customerEmailRequest) {
        logger.info("[Customer] Iniciando criação de cliente por e-mail");
        final User customer = CustomerMapper.toDomain(customerEmailRequest);
        final User customerSaved = this.createCustomerWithNameAndEmailUseCase.execute(customer);
        logger.info("[Customer] Cliente criado com sucesso por e-mail: id={}", customerSaved.getId());
        return ResponseEntity.status(HttpStatus.CREATED).body(customerSaved);
    }

    @GetMapping("/email")
    public ResponseEntity<User> getCustomerByEmail(@RequestBody CustomerByEmailRequest customerEmailRequest) {
        logger.info("[Customer] Buscando cliente por e-mail");
        final User userFound = this.findUserByEmail.execute(customerEmailRequest.email(), customerEmailRequest.userRole());
        logger.info("[Customer] Cliente encontrado por e-mail: {}", userFound.getEmail());
        return ResponseEntity.status(HttpStatus.OK).body(userFound);
    }

    @GetMapping("/cpf/{cpf}")
    public ResponseEntity<User> getCustomerByCpf(@PathVariable @CPF String cpf) {
        logger.info("[Customer] Buscando cliente por CPF");
        final User userFound = this.findUserByCpfUseCase.execute(cpf);
        logger.info("[Customer] Cliente encontrado por CPF: id={}", userFound.getId());
        return ResponseEntity.status(HttpStatus.OK).body(userFound);
    }
}