# src/main/java/com/postech/fastfood/infrastructure/controller/PaymentController.java
package com.postech.fastfood.infrastructure.controller;

import com.postech.fastfood.application.gateways.LoggerPort;
import com.postech.fastfood.application.usecases.interfaces.GenerateQrCodePaymentUseCase;
import java.util.UUID;
import com.postech.fastfood.application.usecases.interfaces.payment.CheckPaymentStatusUseCase;
import jakarta.validation.constraints.NotNull;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/payment")
public class PaymentController {

    private final GenerateQrCodePaymentUseCase generateQrCodePaymentUseCase;
    private final CheckPaymentStatusUseCase checkPaymentStatus;
    private final LoggerPort logger;

    public PaymentController(
            GenerateQrCodePaymentUseCase generateQrCodePaymentUseCase,
            CheckPaymentStatusUseCase checkPaymentStatus,
            LoggerPort logger) {
        this.generateQrCodePaymentUseCase = generateQrCodePaymentUseCase;
        this.checkPaymentStatus = checkPaymentStatus;
        this.logger = logger;
    }


    @PostMapping("/generate-qr-code")
    public ResponseEntity<String> generateQrCode(@RequestParam("orderId") @NotNull UUID orderId) {
        logger.info("[Payment] Iniciando geração de QR Code para o pedido id={}", orderId);
        final String qrCode = generateQrCodePaymentUseCase.execute(orderId);
        logger.info("[Payment] QR Code gerado com sucesso para o pedido id={}", orderId);
        return ResponseEntity.ok(qrCode);
    }

    @GetMapping("/check-status")
    public ResponseEntity<String> checkPaymentStatus(@RequestParam("orderId") @NotNull UUID orderId) {
        logger.info("[Payment] Verificando status de pagamento para o pedido id={}", orderId);
        final String status = checkPaymentStatus.execute(orderId);
        logger.info("[Payment] Status de pagamento verificado para o pedido id={}: {}", orderId, status);
        return ResponseEntity.ok(status);
    }

}