# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: API FastFood
  version: 1.0.0
  description: API para gerenciamento de pedidos, produtos e clientes de um sistema de fast food
  contact:
    name: Suporte Técnico
    email: suporte@fastfood.com
servers:
  - url: http://localhost:8080
    description: Ambiente de desenvolvimento local
  - url: https://api.fastfood.com
    description: Ambiente de produção
security:
  - bearerAuth: []
paths:
...
  /payment/generate-qr-code:
    post:
      tags:
        - Pagamentos
      summary: Gera um QR Code para pagamento de um pedido
      operationId: generateQrCode
      parameters:
        - name: orderId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /payment/check-status:
    get:
      tags:
        - Pagamentos
      summary: Verifica o status de pagamento de um pedido
      operationId: checkPaymentStatus
      parameters:
        - name: orderId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
  /customer/email:
    post:
      tags:
        - Clientes
      summary: Criar cliente com email
      description: Cria um novo cliente usando email e nome
      operationId: createCustomerWithEmail
      requestBody:
        description: Dados do cliente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerEmailRequest'
            example:
              email: cliente@exemplo.com
              name: Maria Souza
        required: true
      responses:
        '200':
          description: Cliente criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    get:
      tags:
        - Clientes
      summary: Buscar cliente por email
      description: Retorna um cliente pelo email (requer autenticação)
      operationId: getCustomerByEmail
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Cliente não encontrado
  /customer/cpf:
    post:
      tags:
        - Clientes
      summary: Criar cliente com CPF
      description: Cria um novo cliente usando CPF
      operationId: createCustomerWithCpf
      requestBody:
        description: CPF do cliente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCpfRequest'
            example:
              cpf: '12345678900'
        required: true
      responses:
        '200':
          description: Cliente criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /customer/cpf/{cpf}:
    get:
      tags:
        - Clientes
      summary: Buscar cliente por CPF
      description: Retorna um cliente pelo CPF (requer autenticação)
      operationId: getCustomerByCpf
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: path
          description: CPF do cliente
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Cliente não encontrado
...
components:
  securitySchemes:
    bearerAuth:
      type: HTTP
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido nos endpoints de autenticação
  schemas:
...
    User:
      type: object
      description: Representação genérica de um usuário (cliente ou funcionário)
      properties:
        id:
          type: string
          format: uuid
          description: ID único do usuário
        name:
          type: string
          description: Nome completo
        email:
          type: string
          description: Email
        cpf:
          type: string
          description: CPF (apenas números)
        role:
          type: string
          enum:
            - ROLE_CUSTOMER
            - ROLE_ADMIN
            - ROLE_MANAGER
            - ROLE_KITCHEN
            - ROLE_DELIVERY
            - ROLE_GUEST
          description: Perfil de acesso
        createdAt:
          type: string
          format: date-time
          description: Data de cadastro
        updatedAt:
          type: string
          format: date-time
          description: Data da última atualização
    CustomerEmailRequest:
      type: object
      description: Dados para criação de cliente com email
      properties:
        email:
          type: string
          minLength: 1
          description: Email válido
        name:
          type: string
          description: Nome completo
      required:
        - email
        - name
    CustomerCpfRequest:
      type: object
      description: Dados para criação de cliente com CPF
      properties:
        cpf:
          type: string
          description: CPF (apenas números)
      required:
        - cpf
...